Updated basic Cove application in Qt.
This is an updated version of our Cove application called Cove Client.
This updated client adds the chat room window and fixed the
functionality of various window buttons. This is still a basic client,
the functionality of which will be implemented later. *This client will
only run in Qt Creator. Written by Beau B.